requirements.txt
streamlit
pandas
Pillow
scikit-learn==0.24.2